Use an action plugin instead of the simple img.php
The problem was that the rendered ditaa file disappeared with the cache
and was not re-rendered when accessed. Now we bind the ditaa to the
current page, via metadata, and access it via the md5 hash. This way we
can easily create the image without knowing if the page still has a
valid cache file for the ditaa (we simply recreate that as needed).